Fault-Plane Automation Determination [A Machine Learning (ML) Based Algorithm]: For Either Seismological/Earthquake or Microseismic Data, “Swarm-Cloud” Investigations
We aim, through the use of this Colab/Python algortihm, to determine the location of the defaulted plane within a given event swarm in order to create a product that companies can use to map the location of previous seismic events to reference a possible fault and corresponding directionality. First be sure to place the csv file in the folder structure on the left of the screen.